+++ flywire_review/__init__.py
@@ -3,4 +3,4 @@
 from .model import Bundle, CodeUnit, NeuronAnnotation, RouteResult
 
 __all__ = ["Bundle", "CodeUnit", "NeuronAnnotation", "RouteResult"]
-__version__ = "0.2.0"
+__version__ = "0.3.0"